--- readme.es.md
@@ -1,27 +1,0 @@
-## GoIT Node.js Course Template Homework
-
-Realice un fork de este repositorio para realizar las tareas (2-6).El fork creará un repositorio en su
-
-[GitHub: Let’s build from here](https://github.com/)
-
-Agregue al mentor como colaborador.Para cada tarea, cree su propia rama.
-
-- hw02
-- hw03
-- hw04
-- hw05
-- hw06
-
-Cada nueva rama para la tarea debe crearse desde la rama master.
-Después de completar una tarea en su rama, debe hacer una solicitud de extracción (PR). Luego, agregue al mentor para la revisión del código. Solo después de que el mentor apruebe el PR, puede fusionar la rama de la tarea en la rama principal.
-
-Lea atentamente los comentarios del mentor. Corrija los comentarios y haga un commit en la rama de la tarea. Los cambios se reflejarán automáticamente en el PR después de que envíe el commit con las correcciones a GitHub. Luego, agregue nuevamente al mentor para la revisión del código.- Al entregar una tarea, incluya un enlace al PR.
-
-- El código JavaScript debe estar limpio y claro, y se debe usar Prettier para el formato.
-
-### Comandos:
-
-- `npm start` — Inicia el servidor en modo de producción.
-- `npm run start:dev` — Inicia el servidor en modo de desarrollo.
-- `npm run lint` — Ejecuta la comprobación del código con eslint. Debe ejecutarse antes de cada PR y corregir todos los errores del linter.
-- `npm lint:fix` — Lo mismo que la comprobación del linter, pero con correcciones automáticas de errores simples.
