+++ routes/api/contacts.ts
@@ -1,0 +1,23 @@
+import { Router } from 'express'
+
+export const contactsRouter = Router()
+
+contactsRouter.get('/', async (req, res, next) => {
+  res.json({ message: 'template message' })
+})
+
+contactsRouter.get('/:contactId', async (req, res, next) => {
+  res.json({ message: 'template message' })
+})
+
+contactsRouter.post('/', async (req, res, next) => {
+  res.json({ message: 'template message' })
+})
+
+contactsRouter.delete('/:contactId', async (req, res, next) => {
+  res.json({ message: 'template message' })
+})
+
+contactsRouter.put('/:contactId', async (req, res, next) => {
+  res.json({ message: 'template message' })
+})
