--- server.js
@@ -1,5 +1,0 @@
-const app = require('./app')
-
-app.listen(3000, () => {
-  console.log("Server running. Use our API on port: 3000")
-})
