+++ server.ts
@@ -1,0 +1,5 @@
+import { app } from './app'
+
+app.listen(3000, () => {
+  console.log('Server running. Use our API on port: 3000')
+})
